Welcome Scene / Fonts
Were added a  welcome scene for the moment to open appication called Bienvenida, also a script for manage the time of view from the scene, called timer.
The location from the scene is in Assets/Scenes/ directory, and the location from the script is in Assets/Scripts/ directory.
Also was changed the font type of the application.